% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_creg.R
\name{AIC.creg}
\alias{AIC.creg}
\title{Generic AIC function for class creg}
\usage{
\method{AIC}{creg}(x, k = 2, ...)
}
\arguments{
\item{x}{A creg object}

\item{k}{Penalty term (default is 2)}

\item{...}{optionally more fitted model objects.}
}
\description{
This function implements the generic AIC function for the S3 class
creg
}
