% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_creg.R
\name{BIC.creg}
\alias{BIC.creg}
\title{@title Generic BIC function for class creg}
\usage{
\method{BIC}{creg}(object, ...)
}
\arguments{
\item{object}{A creg object}

\item{...}{optionally more fitted model objects.}
}
\description{
This function implements the generic BIC function for the S3 class
creg
}
