% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_creg.R
\name{creg_calc}
\alias{creg_calc}
\title{Function to calculate the likelihood of a copula with two distributions}
\usage{
creg_calc(
  data,
  param,
  dist1 = Norm(),
  dist2 = Norm(),
  copula = "Frank",
  param_dim
)
}
\arguments{
\item{data}{Input for the two distributions}

\item{param}{A matrix/data.frame which defines the parameters for the
distributions and the copula.}

\item{dist1}{First distribution, see Details}

\item{dist2}{Second distribution, see Details}

\item{copula}{Defining the copula for the calculation. Can either be "Frank",
"Gumbel", "Gaussian" or "Clayton".}

\item{param_dim}{To correctly extract the parameters from the parameter
matrix, the number of parameters per distribution has to be defined. This
is done by providing a vector with two parameters. E.g. c(2,1)}
}
\value{
Depending on the choice of probab, either a likelihood or the result
of the copula is returned.
}
\description{
This function calculates the likehood of a copula from two
distributions.
}
\details{
This function is primarily used to return the sum of log-likelihood
for the optimization of the beta coefficients of the global_creg function.
For the exact formulas, please consult the documentation fo global_creg
}
\seealso{
\code{\link{global_creg}},\code{\link{creg_opt}},
}
