% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creg_opt.R
\name{creg_opt}
\alias{creg_opt}
\title{Optimization function for global_creg}
\usage{
creg_opt(
  creg,
  Z = NULL,
  data = NULL,
  dist1 = NULL,
  dist2 = NULL,
  copula = NULL,
  startbeta = NULL,
  param_trans = NULL,
  method = "L-BFGS-B",
  cores = 1,
  maxit = 2e+05,
  infinity_control = NULL,
  recalculate = FALSE
)
}
\arguments{
\item{creg}{Creg object from global_creg.}

\item{Z}{Alternative input for the parameter calcualtion,
if no creg object is provided}

\item{data}{Alternative data for the optimization of beta factors,
if no creg object is provided}

\item{dist1}{If no creg object is provided, the first distribution has to be
defined.}

\item{dist2}{If no creg object is provided, the second distribution has to be
defined.}

\item{copula}{If no creg object is provided, the copula for the calculation
needs to be defined}

\item{startbeta}{Optional starting value for the betas can be supplied.
If no value is provided, the result of the creg object is used. If no creg
object is provided either, the starting values are set to 1.}

\item{param_trans}{If no creg opject is provided, tt is possible to assign
a transformation function, which transforms the parameters to a specific
format (e.g. circular,binary,...)}

\item{method}{Method for the optimization. Note that for parallel
optimization only the "L-BFGS-B" method is supported. For a single core
optimizaiton other options are: "Nelder-Mead", "BFGS", "CG", "SANN","Brent"}

\item{cores}{Defines the number of cores that are used for the optimization.}

\item{maxit}{Defines the maximal number of iterations.}

\item{infinity_control}{Optional parameter. See details}

\item{recalculate}{Logical parameter, whether the optimal model should be
recalculated and returned instead of the optimization results. Default is
FALSE}
}
\value{
Returns the optimizing results from the optimize function.
}
\description{
This function is a short cut for the optimization of the
global_creg function. The optimisation is either performed by optim() or by
optimParallel() of the optimParallel package.
}
\details{
This is the optimization function, to optimize a creg object from
the function global_creg. Within this function both one core optimization
with optim() as well as multiple optimization through optimParallel() can
be computed.


If no creg object has been calculated, it is optionally possible to provide
the corresponding variables instead.
Note that the format of the function has to be in accordance with the +
global_creg definitions.

The parameter infinity_control can be used to assign a value, in case the
Log-Likelihood of the global_creg function is returned as Inf or -Inf. In
case of the "L-BFGS-B" method the infinity_control value is per default
10e50 (or -10e50 for negative Inf respectively).
}
\seealso{
\code{\link{global_creg}}
}
