% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DvonMises.R
\name{DvonmisesRad}
\alias{DvonmisesRad}
\title{Von Mises distribution function}
\usage{
DvonmisesRad(x, mu, kappa)
}
\arguments{
\item{x}{Input parameter}

\item{mu}{Mu parameter}

\item{kappa}{Kappa parameter}
}
\value{
Returns density
}
\description{
This functions returns the density  of the
von Mises distribution from the Circular package
}
