% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assisting_functions.R
\name{global_dist_calc}
\alias{global_dist_calc}
\title{Assisting function for the creg_calc function}
\usage{
global_dist_calc(Y, param, k, dist)
}
\arguments{
\item{Y}{Vector, Input for the distribution}

\item{param}{Parameter of the distribution. These can range from 1-5
parameters.}

\item{k}{Number of parameters of the distribution}

\item{dist}{Distribution. This can either be a list, vonMises or from the
dist() package}
}
\value{
Returns a matrix: first collumn is the probability, second the
density
}
\description{
This function calculates the density and probabilty of a given
distribution
}
\details{
See the documentation of global_creg for more information
}
