% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{helgoland_wind}
\alias{helgoland_wind}
\title{The following dataset is available in the AddModCopula}
\description{
The dataset contains the hourly observations of wind speed and
wind direction of the island Helgoland. The collumns are defined as followed:

ws: wind speed in m/s
wd: wind direction, ranging from 0 to 2*pi
dwd: describes the derivation of wd (see example for the code)
}
\examples{
calc.dwd <- function(wd) {
dwd <- rep(0, length(wd))
minvec <- c(0, 0, 0)

for (i in 2 : length(wd)){
 # Version for 2 pi scale
 minvec[1] <- wd[i] - wd[i-1]
 minvec[2] <- (wd[i] + 2 * pi) - wd[i-1]
 minvec[3] <- (wd[i] - 2 * pi) - wd[i-1]
 dwd[i] <- minvec[min(abs(minvec)) == abs(minvec)]
 }
return(dwd)
}

}
\references{
\url{https://www.dwd.de/}
}
\author{
Malte Lehna
}
\keyword{data}
