% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_creg.R
\name{infcrit}
\alias{infcrit}
\title{Helper Function of AIC and BIC}
\usage{
infcrit(object, BIC, k, ...)
}
\arguments{
\item{object}{A creg object}

\item{BIC}{Optional Parameters if BIC is required}

\item{k}{Penalty terms}

\item{...}{Additional parameters}
}
\description{
This function implements the generic AIC and BIC
for the S3 class creg
}
\details{
It is possible to receive both the AIC and the BIC from the creg
object, by using the AIC function. An additional parameter (with
BIC=TRUE) can be passed into the function to receive the BIC value.
}
