% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_creg.R
\name{plotting}
\alias{plotting}
\title{Assisting Plot function for class creg}
\usage{
plotting(x, option, ...)
}
\arguments{
\item{x}{creg object}

\item{option}{Object to define the respective plot.}

\item{...}{optionally more fitted model objects.}
}
\description{
This function implements the generic plot functions for the
S3 class creg
}
\details{
The next generic function is the plot() function, which uses ggplots to
visualize the results. Similar to other generic plot functions, it is
possible to recieve different plots, by including the additional parameter
option e.g. \code{plot(x,option="likelihood3D")}.#'
The following plots are available within the function:
"likelihood2D" and "likelihood3D" display the individual likelihood of each
observation in a 2D or 3D plot. Note that the value of the likelihood
corresponds with the result of the joint density.
In addition, it is possible to plot the results of the copula density by
setting option="copula3D". If option="dataview" is selected, three plots for the
dataset are plotted. The default selection for option is "likelihood2D".
}
