% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PvonMises.R
\name{PvonmisesRad}
\alias{PvonmisesRad}
\title{Von Mises distribution function}
\usage{
PvonmisesRad(q, mu, kappa, tol = 1e-20)
}
\arguments{
\item{q}{Input parameter}

\item{mu}{Mu parameter}

\item{kappa}{Kappa parameter}

\item{tol}{Tollerance level for the probability calculation}
}
\value{
Returns probability value
}
\description{
This functions returns the probability of the
von Mises distribution. Based on the circular package
}
