% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_creg.R
\name{print.creg}
\alias{print.creg}
\title{Print function for class creg}
\usage{
\method{print}{creg}(x, ...)
}
\arguments{
\item{x}{creg object}

\item{...}{Optional parameters}
}
\description{
This function implements the generic print functions for the
S3 class creg.
}
\details{
The print() function returns the Log-likelihood of the
 copula regression
}
