% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_creg.R
\name{summary.creg}
\alias{summary.creg}
\title{Summary function for class creg}
\usage{
\method{summary}{creg}(x, param = FALSE)
}
\arguments{
\item{x}{creg object}

\item{param}{Logical, can be used to display the parameters of the
distribution.}
}
\description{
This function implements the generic summary functions for the
S3 class creg
}
\details{
Through the summary() function  basic details are returned.
 If wanted, additional information can be collected through including the
 paramter param=TRUE. Accordingly,  the summary function will also
 return the parameters of the distributions and the copula, both prior and
 after the transformation.
 \code{summary(x,param=TRUE)}
 @export
}
