% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_creg.R
\name{summary.creg}
\alias{summary.creg}
\title{Summary function for class creg}
\usage{
\method{summary}{creg}(object, ...)
}
\arguments{
\item{object}{creg object}

\item{...}{Optional parameters e.g. param=TRUE}
}
\description{
This function implements the generic summary functions for the
S3 class creg
}
\details{
Through the summary() function  basic details are returned.
 If wanted, additional information can be collected through including the
 parameter param=TRUE. Accordingly,  the summary function will also
 return the parameters of the distributions and the copula, both prior and
 after the transformation.
 \code{summary(object,param=TRUE)}
}
